Updated mediawiki/extensions Project: mediawiki/extensions/Flow  45ae21cab3699b347840f505be104c0ac7ccc489
Convert watch star from wikiglyph glyphs to OOUI icons

The filled star's hover color (bright green) is unique; rather than
creating an additional variant in OOUI specifically for it, I put
in a CSS hack using a brightness filter. There is already a task
for reconsidering the watch star design: T97696.

Depends on Ib92c0644fae in OOUI.

This is now safe to merge, because the dependency has made it to core.

Change-Id: I0540c41c7bdf3f69e171fb2a8b9d674b64059fff