Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  ca088d23d836aa4c1b631a091b1f8a648ad329a7
Support <poem> translations

Adapt the poem (mw:Extension/poem) from the source article to the
target article.

Editing the poem text is not supported. If there is MT, the translation
will be used for target article.

Testplan:
Translate an article with poem in it. Example:
https://en.wikipedia.org/wiki/Puja_%28Hinduism%29
After translating, ensure that translation has poem with content in it.

Bug: T100385
Change-Id: I98d29f889b9afd4497898a2edbfce5366e5e719e