Updated mediawiki/extensions Project: mediawiki/extensions/Gather  e95ecc2938de3849cb8f201d6eddc62b18c1b836
Fix api call for recent/active collections

When renaming the collection to 'active' the api check for the type of
collection wasn't fixed too, leading to pagination errors.

To test this locally go to
http://en.m.wikipedia.beta.wmflabs.org/wiki/Special:Gather/all/recent
by having more than 50 collections with 4 items or more and scroll to the
bottom to see if the second page will load.

It should not fail.

For creating those collections ping me up personally if you can't come up with
a way to do it quickly.

Bug: T103496
Change-Id: Id2f07e12c07c925e35e59c5911112cce1f94e2d8