Updated mediawiki/extensions Project: mediawiki/extensions/SyntaxHighlight_GeSHi  6ba411fe2fa6aebb1b79d73d9e33de8c023d1b28
Merge ext.geshi.data and ext.geshi.visualeditor modules

Until and unless there is an explicit need (and intent to support)
a public interface to use this data elsewhere, there is no need
for it to be a separate module.

It can be its own class, but at the resource level this data should
be considered a JavaScript file. And we don't create separate modules
for each file. They are in the same module unless they should be
semantically loaded in unrelated page contexts.

The module registery is not a javascript class autoloader. There
is a global cost to adding more modules to this registry.

Change-Id: Ifeddef8cfe00b6c115734f92eceab251a0b75bdb