Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  7eefc5fdeb0734835daadf9104fc43975cc6dc0e
Drop leading zeros from TimeValue constructor parameters

These timestamp strings are used to construct TimeValue objects.
TimeValue always ignored leading zeros and does an enforced padding
to 4 digits now.

The puspose of this patch is to make the test cases more readable.
For example, it's very hard to distinguish 0000000000000001 and
0000000000100001 but much easier to distinguis 1 and 100001.

Change-Id: I65f7a6a087c27233fc762083ef98d9fc5e84a382