Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  8a4445f67012a5b55e26a5f66a3e1e80d309d374
Add searchentities alias param only for alias search results

The array and param need to be omitted if the result is not
an alias.

If it is empty array, then the suggester still shows "Also known as:"
and then nothing (no aliases, nothing...).  Previously (and on Wikidata)
"Also known as:" is omitted if there are none.

Bug: T104273
Change-Id: Ia5eaa1ba74a4f6fbc19cef1caf7f7dc42bb70016