Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  3e35ccb8f2323771d2e82649ebcffaaf897eeee1
Remove brackets in include/require_once calls

This patch contains a single change: brackets after "include",
"include_once", "require" and "require_once" are removed.
This is a requirement in the PHPCS rule set we want to use. These
are not functions and should not be used as functions.

I can not enable the relevant PHPCS sniff at the moment. This would
require a much larger patch. I want to avoid that for the sake of
an easy review.

Change-Id: Ibaf0e22d4af435ae52afc79543a07d4c22778688