Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  4f55aa56d5338186cd6dc9d5ac7229ccd0127427
Unify whitespace in array() functions

This is an other step forward to enable more PHPCS sniffs. Again, this
patch focuses on a single thing. Again, I can not enable the relevant
PHPCS sniff at the moment because it would trigger much more errors and
would need a much, much bigger patch. I want this to be done in small
steps for the sake of a much easier review.

Change-Id: I038a0f4830397f058fcdf4e7e5a4266c195351de