Updated mediawiki/extensions Project: mediawiki/extensions/Translate  701f26d92036ebd32ad9af58c5c659df6498db3b
Only hide facets when there are no results at all

The last change hid facets when there were no results in the currently
selected language. When user does a search from the translatewiki.net
main page, the search language is set to the user's interface language.
If the search string was in other language, the user was unable to see
the results because the facets were hidden if there were no matches in
the interface language.

Change-Id: Ie12c34fd18dae57ff73672e772299ece76c84318
Followup-Id: I9e9af0ee30b684884c7b2ef0cec41b76d63ee414