Updated mediawiki/extensions Project: mediawiki/extensions/RandomImageByCategory  029ad1f7a6a37881027fd42289fd5a3dc13a16bd
Change 1 to 0 to work around an edge case when the user-supplied category or categories contain only one image so that we display that instead of nothing

Dev sites and/or brand new wikis don't necessarily have a lot of files
yet, so this is necessary, and I don't think it'll have any adverse
effects on bigger wikis.

Also bumped version number to 1.1, removed now obsolete MEDIAWIKI check
and function-level profiling.

Change-Id: Icd9c073267c14f1f6eb0b281eb393e6485f91c42