Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  464a4fdc3bd1d0b3e9c3b4ae53284489b0e0d59e
Minor cleanups and narrower interfaces in MergeItems code

The reason to touch this was the ApiBase::PARAM_... constants I missed
in Ibeb6d10. I also did a little bit of refactoring and made some
interfaces of private methods smaller: it never can be an Entity, it
always is an Item.

Change-Id: I84952a544c430c9894dd36fa2d7932c5f2d14075