Updated mediawiki/extensions Project: mediawiki/extensions/OAuth  b1b824cb3c98b478694417adc90bcbc1189e06eb
Prevent canonical redirect when OAuth is in use

If the client supplied OAuth headers, we don't want MW redirecting them
to the "canonical" URL, in large part because they're probably not
expecting a redirect. And, at least in Wikimedia's setup, the
redirected-to URL cannot be correctly signed with OAuth anyway.

Follow up to I88ed3525141c765910e66188427b9aab36b958a9

Bug: T105387
Change-Id: I0bc5b35698c97e96fbb91a913c1a0705fbe73bc9