Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  ca93cc9740c8a16325b05301301da7ac7d300510
Reformat dynamic ResourceLoader::register calls for readability

This patch does not change any semantics. It's pure refactoring. I'm
very sorry to say that, but I found the code extremely hard to read
and to understand. This is much more straightforward and much closer to
the existing code in all the other resource.php files. It uses the same
array() structure again. So to a certain degree this is a "revert" of
the formatting done in I4412f14.

Bug: T105211
Change-Id: I123e79fc9e689688b690ea5cd6fa699c873ed777