Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  1cc3327b9f4326253d414a5ccf17eb99042ceefe
Add missing listview dependencies to toolbar controllers

All these modules call methods that are declared in
jquery.wikibase.listview, e.g. listview.listItemAdapter(). The missing
dependency is not an actual problem becaus the dependency is already
loaded via the other dependencies. But it's better to have it explicit.

This is split from I9444ae5.

Bug: T87759
Change-Id: I8bb1301110467a4bed5d172855c940341fb7d57b