Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  e53c8ea477e24d55bd1fe879221a1e649fe9db81
Log an executor id in ElasticsearchIntermediary

Auto forwarding did you mean (I7ed79942) will mean that queries with
no results and requested via the web will be repeated. We need a way
to tie together these two queries in the log files and treat them as
one for the stats.

Patch updates CirrusSearchRequests to include an executor id. This
value is unique to a particular web request or job run and will allow
us to group together queries that happened together. When running
analytics against these logs only the last minute perhaps of
executor ids need to be remembered.

Bug: T105202
Change-Id: I7997c17f9d73d6070c7416b5de21f8b1fe8a5cc4