Updated mediawiki/extensions Project: mediawiki/extensions/Flow  cceab1c6b6d6e613ca99925f53f4286aeefcbb9c
Force scrollbars on OS X

OS X has a definition to hide scrollbars by default. This is an
unwanted behavior specifically in the ToC because then the user will
have trouble scrolling to the bottom of the list. This fix is
adding special webkit rules to re-enable (and rebuild) those scrollbars
even if they're hidden.

Bug: T106029
Change-Id: I938a5ec58f897cd654c79e023f4132e5e797510c