Updated mediawiki/extensions Project: mediawiki/extensions/SocialProfile  cf0c2f84af6585a6d49354861e61242cb8de031a
If the birthday is empty (0000-00-00), don't incorrectly format it and show it on user profiles against their wishes.

Previously 0000-00-00 would be formatted as "November 30th", no matter
what, and shown on user profiles. This was incorrect on so many different
levels.

The code here is still kinda nasty and should be refactored to use the
proper Language method(s), but at least this fixes that particular bug.

Bug: T88912
Change-Id: Ia37b5969fbfe9e64181816c26c3e8f0a8243a2cd