Updated mediawiki/extensions Project: mediawiki/extensions/Gadgets  e858bda3088da7090f01d6c8c670921fc4e8ff7e
Add abstract GadgetRepo and MediaWikiGadgetsDefinition implementation

GadgetRepo is an abstract class based off of the implementation in the
RL2 branch.

It is a singleton that provides basic methods to construct and interact
with Gadget objects.

The MediaWikiGadgetsDefinition class is an implementation of GadgetsRepo
that parses the "MediaWiki:Gadgets-definition" page for gadget
definitions.

Tests were left in place to demonstrate that no functional changes have
been made aside from relocation of code. Some tests should be moved to
separate files in the future.

Bug: T106176
Change-Id: I3e802889f6f495783f4dbac65c2a8cefa824a778