Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  92b1030bd7bd0d70a92928c1d17fe1cf87824b67
Add test for RawMode addClaim(s)? in ResultBuilder

The next patch will mode these methods over to the
DM serializers. This test has been intruduced to make
sure that the api output does not change.

Once all of this refactoring is done we will likely make
a single 'big'ish breaking change to the api output
to enable us to clean all of this up

Change-Id: I78a0dbe1aef11fc36ab531ee90db52e2dbe15610