Updated mediawiki/extensions Project: mediawiki/extensions/TimedMediaHandler  16a636cb8fdf85677c3997aff501f8abaa8c483c
Fix for double-playback of videos in IE/Edge with WebM components

IE/Edge seem to trigger the video's autoplay on removal from DOM
as well as addition to DOM. What?

This fix prevents a second copy of the video from playing in the
background (with overlapped audio) every time you trigger playback
of a popup video transform.

Bug: T106957
Change-Id: Ib4dd5a81cff91faea98484fb7141ecfa8573b140