Updated mediawiki/extensions Project: mediawiki/extensions/MwEmbedSupport  cb17279b82dd1e08ee954b75e12a9fef627f68c7
Fix for fullscreen video in IE 11

IE 11's prefixed fullscreen API is slightly differently
named than other browsers and didn't get detected, leading
to faux full screen done by expanding the video to the
size of the viewport.

Updated to accept both the older names in Gecko etc and the
newer names in IE 11 and natively in Edge.

Bug: T106979
Change-Id: I57bb9c1ff18e997f790a44d519c285a19e6fa048