Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  1958d70b43e5b202ebd50dd59817c3f1f47454ae
Don't swap link cards when opening the link of the card behind

Link cards are presented usually in pairs. Although the card in front
represents the main action (add/remove link in the translation),
it is possible to explore the card behind by either hovering
(the card behind is temporary revealed allowing to click on the link
to open it) or clicking on the card (cards switch positions).
The problem is that when the link from the card behind it is opened,
the click event propagates to the whole card and the cards are also swapped.
Cards should be only swapped when clicking on the card behind except when clicking on the link.

Bug: T106595
Change-Id: I0873eb77b082c06004718a6e7b737d1a9fa97062