Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  18ba3017e7cbfa2ea1e51de5fe9e436d365d41ce
When link target changed, reset the style classes

When the link inspector is used on a red link or a red link to create
a regular link, it is expected for the resulting link to become blue.
Currently, the link remains red or grey (as it was before) which is confusing.

Bug: T106437
Change-Id: Id86d4f154b1fbbbd9c411f24fadc7a7356a8f7fa