Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  f93673cb75ef78414998c19f5dad49eb8d27a8e4
Log unrecognized AstroPay errors at 'error' level

If the error is unfamiliar enough we don't have a user message for
it, we probably want to see it in the 'error' logs, not just warning.

It'll duplicate these messages at both warning and error, but there
should not be too many of them.

Bug: T106053
Change-Id: I4d231743741b35209d50bde0205697230eec4b9a