Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  2bd7b2bd2779250474c99d7b9fc67afeb18b7b1b
Allow disabling cirrus request logging from query string

We want to be able to try throwing various combinations of search parameters at
the production elasticsearch instances for indexes that are too big to test
elsewhere (such as enwiki).

This allows a web request to recuse itself from the CirrusSearchRequests log if
it knows the secret key. The key will need to be set from the cluster's private
data repository.

Change-Id: Iac79ad90e3ae32aa10a8eba5df63e5906ea41aa3