Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  e891be3c29d490a9856ad9b9e40942281ea7e57a
Fix crash in ContentTranslationHooks::onSaveOptions()

It is not safe to chain $wgOut->getTitle()->isSpecial(), because
OutputPage::getTitle() can return null, and the code will crash.
This has occurred in production. Fix it by checking that $title is truthy.

While I'm here, clarify the flow of control by rephrasing some of the
conditionals.

Change-Id: I81b1d021d3cb5fe585c99b29192fd399bf841189