Updated mediawiki/extensions Project: mediawiki/extensions/GoogleLogin  a61ffd0c27e26b1dc219b5861806366c9550a23c
Silently uppercase the first letter of a users name

Rewrite GoogleLogin::isValidUserName() to use User::newFromName with validity
check for a createable username (which takes care of all possible things).

Add some basic tests, but remember, that the main tests are run by MediaWiki core
in UserTest::testGetCanonicalName(), which is the base of the rewritten function.

Extra points:
 * Add right message for managegooglelogin

Bug: T107806
Change-Id: I49c5ba91dd2def979a25e1debc1369621c3d8043