Updated mediawiki/extensions Project: mediawiki/extensions/Teahouse  07eb7773b2e58bad1b4b7851205a6b1af97ddb48
Break out of endless loop

This is only a partial fix for a problem with the current behaviour
of the API. The edit token is not accepted. This creates an endless
loop because the script tries the same request again and again with
the same token. A simple !== null check stops the loop after the
second try (the first was without a token, the second with a token).

Change-Id: I424e60bdbde19794798033c44547001dc4a9985a