Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  0f2030f00eb7b5f67f158b4b5c4dbcc9ec832c32
Skip tests that have side effects

These tests all stub inside the test itself rather than setup and teardown.
This is bad practice and seems to unintentionally break other tests.
I will personally go through these and get them turned on but right now I want
to run all the tests that are behaving that are being excluded by Jenkins because of
this.

Bug: T104362
Change-Id: I588a9180f903842e5ef53aa34daa348ff1463adf