Updated mediawiki/extensions Project: mediawiki/extensions/TimedMediaHandler  0821cf6ddab8414a5b13670088ea70a86447b398
Mark parseroutput that includes TMH elements

In order to be able to fix T74675, let's mark all parseroutput with a
flag if TMH is included inside it. After 30 days, all cached output
with TMH enabled will then be recognizable from a
RejectParserCacheValue and we can start making RL module changes.

Change-Id: Ia36cd3e5ec2514276f1205b0e5d31742e4794dc7