Updated mediawiki/extensions Project: mediawiki/extensions/LinkFilter  70abb0a588313768d4f8a48753baaa7df3982f1d
Version 3.2.1: client-side checks to ensure that there is a link URL and description

These have always been enforced in the PHP files for
Special:LinkSubmit/Special:LinkEdit.
Enforcing them here means we offer a better user experience for mobile
users in the case that they forget to fill in one (or more) of the
mandatory fields. Previously if there was no URL or description, the JS
would not alert the user and would happily submit the form (which would
then proceed to fail the PHP-side sanity checks).

Change-Id: Icafdfad3f4fc0f64d0f297102ba011607775d407