Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  e0319f28a1de0d771c866fadccfa644d431413ec
Page actions should not be cut off on the right

Do not set a width which you do not control. The width is defined
by the mw-ui-icon class which we have no control over. Instead use
display: block to force the icons to be stacked in mobile and use
inline-block to make them inline on tablet.

Bug: T108641
Change-Id: I7d249823bafeadfd86e89da1a4abef3b77a957d3