Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  c9a7ac69775644585c57a141f8e2bf9e2306bbfe
Change the id of source title so that its prefix is same as parsoid ids

Recently we moved to parsoid generated Ids and our id validations
changed. But the id of source title did not change accordingly and
it broke the connection between source and target title elements

And then it caused breakage of dynamic alignment of titles

In this patch we keep the id prefix as mw.

Testplan:
Start a translation, In the target title, type some text that is so
long that it cause line break. Observe if the titles of source and
translation aligns dynamically.

Bug: T108524
Change-Id: I02a1af3f2b8d00cb54c099f4c91c335d736d69a0