Updated mediawiki/extensions Project: mediawiki/extensions/QuickSurveys  e2d3ff007e1e39cc508632a936ba03539931501f
Revert to JSCS wikimedia preset

The current one doesn't work as it runs on the Gruntfile.
It's also soon to be out of date (see Ifb239a232a2271e757add0315fac6564d9773d59)

Ideally we should not be deviating from the preset and we should be discussing
and upstreaming changes and only overriding where necessary!

Change-Id: I60843d93ff169c38d644cb45208eb9530366d470