Updated mediawiki/extensions Project: mediawiki/extensions/Flow  dca7fe80745d79726537382585e839c5e68a0e48
Make the reply link for the last post work

The reply link for the last post points to the ID of the
topic, not the ID of the post, so it needs to be handled
differently. It also needs to hook into the (initially collapsed)
ReplyWidget that's already present at the bottom of each topic.

Bug: T109143
Change-Id: I26b030656c2a80d7b2508654a02fd8a1805ce375