Updated mediawiki/extensions Project: mediawiki/extensions/CentralAuth  f1088fb68284d3c28bcedfa7529bbec37d141e89
Avoid throwing SecurityError on window.localStorage access

When cookies are disabled in Firefox, merely referencing window.localStorage is
enough to cause Firefox to throw a SecurityError exception. Avoid it by
wrapping localStorage operations in try / catch, as we already do in
mediawiki.js (and as CentralAuth already does in anon-set.js).

Change-Id: I28399cf8b63328ee34316b785f2d091f6f052224