Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  f20737008c1ff01f53294470b01f6f69691e6473
Refactor the dashboard and translationlist modules

Translation list module was handling the rendering and event
handlers of translation status filter and language selectors.
It was also rendering the new translation button.

This commit removes all the above features to the dashboard module.
Translation list is now just a widget to list the translations.

This change is to initiate the listing of suggestion initiated by
dashboard module and provide way for suggestion list to share the
language filters.

Functionality wise nothing changed.

Change-Id: Ie5bbcb3a01dfa0b136ad8fa000ca1f52a7a98878