Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  6db7cec3c9c92bd2cab9bfc113aaddc2cb4c2669
Allow CIDR ranges to be opted into per-user poolcounter

We have a small number of IP's issuing a large number of
queries against the cluster. We originally created this
per-user poolcounter to deal with them but never enabled
it due to uncertainty around proxys.

This patch lets us provide a list of ip's to be opted into
the poolcounter, side stepping the issue about proxies
but allowing us to turn it on in a targeted fashion for
abusers.

Bug: T76497
Change-Id: I434280364c37a0f87688f15a7294cba71b7efb37