Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  060c6aa122142d751ce0779e34e5e138739cbe3c
Refactor the dashboard and translationlist modules

Translation list module was handling the rendering and event
handlers of translation status filter and language selectors.
It was also rendering the new translation button.

This commit moves all the above features to the dashboard module.
Translation list is now just a widget to list the translations.

This change is to initiate the listing of suggestion initiated by
dashboard module and provide way for suggestion list to share the
language filters.

Functionality wise nothing changed.

Change-Id: I647c4b5348e59f3c5385a08907b1953da113952e