Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  7de1a7c5841e47040981d7e74d6a03f6880f983e
Validate the references when restored from draft

If the references in source article changed, the references
in translation might point to invalid items in references list.
This can cause publishing error because of parsoid error.

For such dangling references in translation, remove them if they
point to invalid references in references list.

Bug: T109574
Change-Id: I83702f8c1df5e8a79dc185a34fc03acd93496415