Updated mediawiki/extensions Project: mediawiki/extensions/OpenStackManager  102ae3bae8bbca09d5bccfe03856be5a9734efa3
Don't cache the service catalog for nearly as long.

Default behavior is to cache the OpenStack service catalog for
the lifetime of a user login.  That makes it pretty hard for us
to move services.

This patch changes the lifetime to 1 hour, and also adds logic
to detect an expired catalog (which was previously ignored.)

Task: T109653
Change-Id: If8d27bce5fef89633448bc66fd2a4f19ed5a3f2d