Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  e808762870eeed260f60af2fa3af817ae09e0e44
Make mobile site module position bottom.

Any radical changes to the UI should not be done in Mobile.css/Mobile.js
If those are necessary the software should support them. These styles should
load at the bottom and enhance content but not in such a way that it leads to
a jerky experience.

Bug: T106756
Change-Id: I577bcc7e75a2e75ebec9123ffa4225c939a35b22