Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  7c673d886fc5b7c11a0a5b478cb717061dc4fcec
Add lots of array type hints to tests

The idea is pretty simple: This patch only touches tests. Code coverage
is not an issue in test code, right? So if all these tests still succeed
with the added type hints, they can't be wrong. There can not be an
unknown caller that calls one of these methods with an other type, right?

Change-Id: I16204cd0506e3fee9f72917867c770ad9d9febaa