Updated mediawiki/extensions Project: mediawiki/extensions/Cite  04ac55ed52ecee505677ee151894ee2d85c0f1f9
Revert "Use display: inline-block instead of unicode-bidi"

I think this change causes too many issues on non-LTR wikis
(see the associated tasks) to be acceptable in the current form.

Maybe it should be redone to apply to RTL text only if it's
worth it there (apologies, but I don't quite understand the issue).

This reverts commit e9c1521fcafa09e50c54765d2d6154863bd55a42
and commit 59ef3c8678408657763e1d5c1d040a0864b30313 (which was
the previous attempt to fix it and caused different issues).

Bug: T108493
Bug: T110057
Change-Id: I205400c59a342dcc4cdf89791bc1c761805cd2a8