Updated mediawiki/extensions Project: mediawiki/extensions/WikibaseQualityExternalValidation  e422ed4ffbf4f31eb3467ac7d56bbde304136215
Fix non-callable type in wgAPIModules registration

I can not tell why this started to fail on my local mashine, because
the relevant is_callable check is there since 2014. However, I think
this is a reasonable change that should be done no matter what. A
string is usually not callable, you know?

Change-Id: I82291273a06ca8c43c688ad625f60a689e6a9c27