Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  bfd1f0d174c5789a40827105ab038d7e639e00fd
Partly revert "alphabetic" InterwikiSorter order

This partly reverts patch I1b415cb. The problem is: What if the sort order
array says [x, k, a] and you need to add the unknown value "b"? Where should
it go? There is no alphabetic order the code can follow. I could pick a
"random" position that looks like it follows an alphabetic order, which is
what patch I1b415cb did. That implementation even made it *worse*: The
example results in [a, b, x, k], so "b" is in an acceptable place, but "a" is
misplaced now.

This patch adds more tests and reimplements "unknown goes to the end, in an
alphabetic order".

Bug: T103044
Change-Id: Ib9add179a6f22fddd5a3a3b68d660588e56aa157