Updated mediawiki/extensions Project: mediawiki/extensions/Newsletter  01526dfd9d890d76f568ff22741e263ea8cce3ee
Add getSubtitleLinks to its own class

This is used by all special pages thus should
be somewhere (not in 1 of them).

We cant go down the route of a base class as
the 3 special pages use different basses.

Could be made non static at some point to improve
the testability....

Change-Id: I737d0c4c6c3f999ef0bab0c95b9b81b40e8ca735