Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  215fd211e8300cf47b37ddabcc255a3f5e0f8f8e
Tweak the cancel/refund logic

Try cancelling first, it's the only way to stop a charge that hasn't settled
yet.

Do some sketchy stuff with the status code mapping, we're busting DI's rivets,
we'll have to redesign the way data and state are handled to be more
appropriate for workflows other than normal payment.

Bug: T110367
Change-Id: I8848fe6542f46f8284ff8b05bbab401dd0af35a4