Updated mediawiki/extensions Project: mediawiki/extensions/Flow  dc10381d1fe3506112bbabf374761291afe957d6
Anchor the regex for ./ stripping

So it doesn't strip ./ or ../ out of the middle of page names.
Apparently it's common to have user names on hewiki end with dots,
so their user subpages end up being of the form [[User:Foo./Bar]].
This unanchored regex corrupted links to those pages.

Bug: T110676
Change-Id: I0cb3e47afd2b3d0d9d953f6650f5858f9bdb58e9