Updated mediawiki/extensions Project: mediawiki/extensions/SecurePoll  f6e39c6d71d63040968c51d2df3b10a10d7a2ad0
purgePrivateVoteData: Don't cronspam if table isn't there

We could conceivably check if the extension is enabled or not, but if the
extension was at one point enabled, then the correct thing to do is to probably
purge the data and not skip the wiki. So check for the table's existence
instead.

Change-Id: I66edf739a9594a23bcff80fdcb831497fa8482e6