Updated mediawiki/extensions Project: mediawiki/extensions/PollNY  5448432da6e9ceb429f42f28c025b098c40827d9
Version 3.3.1: don't show the "discuss" link on embedded polls when $wgPollDisplay['comments'] = false;

If a page embeds a poll with the <pollembed> tag and a user has already
voted in the poll, the "Discuss" link would show up irregardless of the
$wgPollDisplay['comments'] setting, which was silly, because commenting on
polls is disabled by default and requires the Comments ext. to be
installed.

Reported on MediaWiki.org by BradLeeBH:
https://www.mediawiki.org/w/index.php?title=Extension_talk:PollNY&oldid=1870956#BUG:_Discuss_textlink_still_shows_on_an_embeded_page_when_using_.3Cpollembed.3E

Change-Id: I0f5561b8db3d54f622d2ec00dffa0897b6be91bb