Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  b268db8ec8ea9b6b6c149e026c6a9abc3c13bf82
Clean up EntityChangeFactory

This fixes an issue I found while reviewing I8f646c1. The type of the
two objects is guaranteed to be the same, that's why checking one of
them is enough.

I also tried to clean up some more code, simplify data flow (e.g.
pass an EntityId instead of an Entity) and such in the same class.

Change-Id: I56dc68c153044c04e5f115025eaa7d696abf95ff