Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  ccc1ce5767b39bbde21528c10f22fb2ae092793b
Consider the object reference not being passed in the expected format

This fix cares for the parameter objref not containing the expected pipe
character as the delimiter. The href of the button that is configured to
'useObjref' will only be applied if the objref parameter consists of at
least two values delimited by the pipe character.

Bug: T111381
Change-Id: I1e422f135aad29d569b81a867da9a393c974f31e