Updated mediawiki/extensions Project: mediawiki/extensions/Flow  0e6b88570d0ed517c3269b7c0f95e71892bf9c17
Add debugging code for invalid titles in WikiReference

This happens when there is an invalid title (e.g. contains '|'), not
just when it's a red link (non-existent).

I'm not sure if we can do anything about it here, but this will help
track down the bad data, and then how it's getting into flow_wiki_ref.

Plus, minor comment fixes.

Bug: T111267
Change-Id: Ie22b02e12aaf27c05f7730d67b105d0b43e22fc0