Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  2bd2a0bd71b5b00050ceb02e8c8a9f742aaa4098
Set space between category name and category close button robustly

margin-left is correctly flipped for RTL
when the target language is the same as the UI language,
which is usually the case, but it might also happen that
it's different, and then it will be incorrect.

So instead of adding space before the close button using style,
insert an element between the text and the button.

To test, set the UI to Hebrew and try to translate something
from English to French.

Change-Id: Idadb6aba8d87fffb933501ecb81c1326f97fbb0f