Updated mediawiki/extensions Project: mediawiki/extensions/QuickSearchLookup  adb2964478bbc47ad13f69314b0ab64602fc15eb
Don't force 100% width

Images with nearly the same width as height would be really really big,
which would look terrible, instead of providing a real feature.

Remove the 100% and use it as a max-width (so images doesn't overflow the
border of the panel). Use a maxheight of 18em for the image to scale.

Change-Id: I3f7c669a49921d3cd528f0b7457eb310dd674942