Updated mediawiki/extensions Project: mediawiki/extensions/Teahouse  489acbb62f44d889e847cd1ea41bde36dcf0f686
Remove duplicate newlines from templates and placeholders

All these newlines result in a page that contains big spaces between each
question. And no way to remove them because the major factor is the "\n\n"
in the code, not the ones in the templates.

The headline is tuned down to a <h2> headline in wiki sytnax to make it
look more like all existing talk pages.

Change-Id: I4d14a7a296354c924d2842193456317f54821f64