Updated mediawiki/extensions Project: mediawiki/extensions/WikibaseQualityExternalValidation  3115fa9188fa3e979b4bd0f351cbec73f2dc95a0
Don't explicitly define wikimedia/assert in composer.json

This is not going to work nicely with our build step,
having multiple copies (one in the build, one in the
vendor directory of core).

Instead we would need our extensions to be installed
with the ordinary process for extensions and libraries
to go in core's vendor directory.

Until then, I don't think we should explicitly define
it and we already depend on core anyway.

Change-Id: I057c022a4816432fbed821d734933a773cb38d70